+++ client-app/src/desktop/common/grid/options/GridDisplayOptions.ts
@@ -170,7 +170,7 @@ export const expandCollapseButton = hoistCmp.factory<ExpandCollapseButtonProps>(
                 } else {
                     gridModel.expandAll();
                 }
-                model.setBindable('expanded', !model.expanded);
+                model.expanded = !model.expanded;
             }
         });
     }